refactor: rename file_handling module to fs for shorter name
