+++ Cargo.toml
@@ -10,15 +10,15 @@ file-handling = []
 
 [[example]]
 name = "read_file"
-required-features = ["file-handling"]
+required-features = ["fs"]
 
 [[example]]
 name = "using_file_paths"
-required-features = ["file-handling"]
+required-features = ["fs"]
 
 [[example]]
 name = "write_html"
-required-features = ["file-handling"]
+required-features = ["fs"]
 
 [dependencies]
 pest = "2.5.5"
