+++ src/lib.rs
@@ -1,5 +1,5 @@
-#[cfg(feature = "file-handling")]
-pub mod file_handling;
+#[cfg(feature = "fs")]
+pub mod fs;
 pub mod html;
 pub mod ytml;
